refactor (navbar): app and navbar components to improve structure and hide scrollbars
